+++ infrastructure/database/models/purchase.py
@@ -1,7 +1,7 @@
 from typing import Optional
 
 from sqlalchemy.orm import Mapped, mapped_column
-from sqlalchemy import Integer, String, ForeignKey
+from sqlalchemy import String, ForeignKey, BigInteger
 
 from infrastructure.database.models.base import Base, TableNameMixin, TimestampMixin
 
@@ -17,8 +17,8 @@ class Purchase(Base, TableNameMixin, TimestampMixin):
         purchase_amount (float): The amount of the purchase.
         purchase_status (str): The status of the purchase.
     """
-    purchase_id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True, nullable=False)
-    user_id: Mapped[int] = mapped_column(Integer, ForeignKey('users.user_id'), nullable=False)
+    purchase_id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True, nullable=False)
+    user_id: Mapped[int] = mapped_column(BigInteger, ForeignKey('users.user_id'), nullable=False)
     username: Mapped[Optional[str]] = mapped_column(String)
     software: Mapped[str] = mapped_column(String)
     status: Mapped[str] = mapped_column(String, server_default="Unpaid")
